wlan: Convert DFS channel back to passive upon timeout
With gActivePassiveChCon enabled, driver converts DFS
channel to ACTIVE when either:
a) the scan channel is the connected channel (or)
b) beacons are received on a DFS channel during scan.
The timer handler is expected to convert the DFS channels
back to passive when there is inactivity on the channel
for 9seconds.

The timer handler works on the last updated timestamps.
However, when the channel is marked as ACTIVE because of
active connection on that channel, the timestamp is not
updated. The timestamp is updated during the reception of
the first beacon. Whereas, in cases when there is no beacon
after the connection, the timestamp never gets marked and
the DFS channel stays as ACTIVE forever.

To fix this, update the timestamp even for scan on connected
channel cases.

Change-Id: I7ea7bb6eff855e512043bc400e0da16d75e54f3a
CRs-Fixed: 3537374