wlan: Convert DFS channel back to passive after disconnect
With gActivePassiveChCon disabled, driver makes current
operating channel as ACTIVE and scans actively. The channel
is expected to be made PASSIVE upon disconnection.

However, in roaming cases due to the presence of session
entry for the new roamed peer in the slot ahead of the
connected peer, wrong channel is made is as PASSIVE.

To fix this, always use the channel info from the session
entry directly to prevent wrong channel getting marked as
PASSIVE.

Change-Id: I9875ba0ff9f9accf4c57bcc10923588e8039afe5
CRs-Fixed: 3537376